fleet-integration/pyproject.toml
[tool.poetry]
name = "fleet-integration"
version = "0.1.0"
description = "Python SDK for integrating devices with the Fleet Controller"
authors = ["Fleet Team"]
packages = [{ include = "fleet_integration" }]

[tool.poetry.dependencies]
python = "^3.11"
paho-mqtt = "^2.0.0"
pydantic = "^2.0.0"

[tool.poetry.dependencies.python-dotenv]
version = "^1.0.0"
optional = true

[tool.poetry.extras]
dotenv = ["python-dotenv"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
